requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
h11==0.14.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
selenium==4.18.1
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.41.2
wsproto==1.2.0
